Add user.* xattr in tmpfs
This patch is an first step to add user.* xattr support to tmpfs.

See https://www.spinics.net/lists/linux-mm/msg109775.html
for a discussion between Pete Zaitcev <zaitcev@redhat.com> and
Hugh Dickins <hughd@google.com> about this topic.

For parisc it would be great to have such support, because we
use tmpfs on the buildd servers, and some packages fail to build
because of missing user.* xattr support on tmpfs.

Signed-off-by: Helge Deller <deller@gmx.de>