parisc: Improve monotonic clock_gettime() resolution on SMP machines
On parisc SMP machines the CR16 CPU-internal clocks are not synchronized
which is why they can't be used to provide a more fine-grained monotonic
clock as it's done on x86 with the TSC timers.
Work around that issue by increasing the standard timer interrupt to fire
10 times more often and use an irq counter to improve the clock_gettime64()
resolution.

Note that this is necessary on physical SMP machines only. Qemu will always
provide a fine-grained clock based on the host clock.

Signed-off-by: Helge Deller <deller@gmx.de>
Cc: stable@vger.kernel.org	# v6.0+
Closes: https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=1020321#23